Try to work around issue #628
On a Linux system with a Vivante GPU, eglGetProcAddress seems
to be implemented via a plain dlsym(), which then resolves the (E)GL
entry points into apitrace itself (and not into the native EGL
libraries). This causes endless recursion into apitrace:

* the traced application wants to call eglXXX
* it calls apitrace's LD_PRELOAD'ed version
* apitrace tries to resolve the real eglXXX, to call it for real (after
logging the call)
* the resolution finds apitrace's version itself
* endless recursion

The "fix" so far is to offer our own version of eglGetProcAddress
that calls the native one, but then checks if the address returned
by it belongs to apitrace or not. If it does, then that result
is discarded and a plain `dlsym` lookup is done.

Note: IT DOES NOT WORK YET.